Server/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
ffmpeg-python
indic-transliteration
aiofiles
pydantic-settings
pyinstaller
